+++ realgui/3rd/box2d/src/dynamics/b2_pulley_joint.cpp
@@ -335,13 +335,15 @@ void b2PulleyJoint::Dump()
     b2Dump("  jd.bodyA = bodies[%d];\n", indexA);
     b2Dump("  jd.bodyB = bodies[%d];\n", indexB);
     b2Dump("  jd.collideConnected = bool(%d);\n", m_collideConnected);
-    b2Dump("  jd.groundAnchorA.Set(%.9g, %.9g);\n", m_groundAnchorA.x, m_groundAnchorA.y);
-    b2Dump("  jd.groundAnchorB.Set(%.9g, %.9g);\n", m_groundAnchorB.x, m_groundAnchorB.y);
-    b2Dump("  jd.localAnchorA.Set(%.9g, %.9g);\n", m_localAnchorA.x, m_localAnchorA.y);
-    b2Dump("  jd.localAnchorB.Set(%.9g, %.9g);\n", m_localAnchorB.x, m_localAnchorB.y);
-    b2Dump("  jd.lengthA = %.9g;\n", m_lengthA);
-    b2Dump("  jd.lengthB = %.9g;\n", m_lengthB);
-    b2Dump("  jd.ratio = %.9g;\n", m_ratio);
+    b2Dump("  jd.groundAnchorA.Set(%.9g, %.9g);\n", (double)m_groundAnchorA.x,
+           (double)m_groundAnchorA.y);
+    b2Dump("  jd.groundAnchorB.Set(%.9g, %.9g);\n", (double)m_groundAnchorB.x,
+           (double)m_groundAnchorB.y);
+    b2Dump("  jd.localAnchorA.Set(%.9g, %.9g);\n", (double)m_localAnchorA.x, (double)m_localAnchorA.y);
+    b2Dump("  jd.localAnchorB.Set(%.9g, %.9g);\n", (double)m_localAnchorB.x, (double)m_localAnchorB.y);
+    b2Dump("  jd.lengthA = %.9g;\n", (double)m_lengthA);
+    b2Dump("  jd.lengthB = %.9g;\n", (double)m_lengthB);
+    b2Dump("  jd.ratio = %.9g;\n", (double)m_ratio);
     b2Dump("  joints[%d] = m_world->CreateJoint(&jd);\n", m_index);
 }
